pyproject.toml
[project]
name = "toy-translator"
version = "0.1.0"
description = "OCR and translation system for Japanese toy instruction manuals using Google Cloud Document AI and Temporal"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
authors = [
    {name = "tokuSolutions"}
]
keywords = ["ocr", "translation", "document-ai", "temporal", "japanese", "pdf"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.12",
    "Topic :: Text Processing :: Linguistic",
]
dependencies = [
    "click>=8.3.1",
    "ftfy>=6.1.0",
    "google-cloud-documentai>=3.7.0",
    "google-cloud-translate>=3.23.0",
    "google-generativeai>=0.8.0",
    "pillow>=12.0.0",
    "pydantic>=2.10.0",
    "pymupdf>=1.26.6",
    "python-dotenv>=1.2.1",
    "requests>=2.32.0",
    "temporalio>=1.20.0",
]

[dependency-groups]
dev = [
    "black>=25.11.0",
    "playwright>=1.57.0",
]
